modified:   main.py Add:        startNewRound()
This function is responsible for generating a new word to guess and
displaying it in the center of the screen as asterisks.

The function starts by generating a random word from a list of words.
This word is stored in the global variable wordComp. Next, the function
creates a string of asterisks of the same length as the word, and stores
it in the global variable wordStar.

The function then sets the text of the wordLabel to the asterisk string,
and sets the text of the scoreLabel to the same string. It also centers
the wordLabel on the screen using the place() method.

The function ends by setting a few global variables to prepare for the
next round of the game.