modified:   main.py Add: pressLetter() getWordStart() compareWord()
The pressLetter function is called when a button is clicked in the GUI,
and it updates the state of the buttons, the wordStar variable, and the
wordLabel to reflect the user's guess.

getWordStart is a function that takes a single character as input
and replaces all the characters in a word that are not equal to the
input character with the same character in the word. The function works
by iterating over each character in the word and comparing it to the
input character. If the characters match, they are added to the return
string. If the characters do not match, the corresponding character from
the wordStar string is added to the return string. The function returns
the modified word as a string.

compareWord is used to compare the user's guess (s1) to the correct word
(s2) in the game.