modified: main.py add:      pressKey()
The function then calculates the index of the pressed button based on
its ASCII value, which is calculated by subtracting the ASCII value of
the first letter ('А') from the ASCII value of the pressed character.
It then checks if the calculated index is within the range of 0 to 32.
If it is, the function calls the pressLetter function with the calculated
index as input.